Move spec stuff to inside service
