+++ Civi/Service/Spec/SpecGatherer.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Civi\API\V4\Spec;
+namespace Civi\API\V4\Service\Spec;
 
-use Civi\API\V4\Spec\Provider\SpecProviderInterface;
+use Civi\API\V4\Service\Spec\Provider\SpecProviderInterface;
 
 class SpecGatherer {
