+++ services.xml
@@ -4,7 +4,7 @@
 
     <services>
 
-        <service id="spec_gatherer" class="Civi\API\V4\Spec\SpecGatherer"/>
+        <service id="spec_gatherer" class="Civi\API\V4\Service\Spec\SpecGatherer"/>
 
         <service id="schema_map_builder" class="Civi\API\V4\Service\Schema\SchemaMapBuilder" public="false">
             <argument type="service" id="dispatcher" />
@@ -21,7 +21,7 @@
             <tag name="event_subscriber"/>
         </service>
 
-        <service id="contact_creation.spec_provider" class="Civi\API\V4\Spec\Provider\ContactCreationSpecProvider">
+        <service id="contact_creation.spec_provider" class="Civi\API\V4\Service\Spec\Provider\ContactCreationSpecProvider">
             <tag name="spec_provider"/>
         </service>
