+++ tests/phpunit/Spec/SpecFormatterTest.php
@@ -2,10 +2,10 @@
 
 namespace Civi\Test\API\V4\Spec;
 
-use Civi\API\V4\Spec\CustomFieldSpec;
-use Civi\API\V4\Spec\FieldSpec;
-use Civi\API\V4\Spec\RequestSpec;
-use Civi\API\V4\Spec\SpecFormatter;
+use Civi\API\V4\Service\Spec\CustomFieldSpec;
+use Civi\API\V4\Service\Spec\FieldSpec;
+use Civi\API\V4\Service\Spec\RequestSpec;
+use Civi\API\V4\Service\Spec\SpecFormatter;
 use Civi\Test\API\V4\UnitTestCase;
 
 /**
